requirements.txt
requests
django
django_restframework
django-celery-beat
django-rest-swagger
Celery
pika
elasticsearch
jsonpath-ng
pyyaml
whitenoise
daphne
textblob
spacy
natsort
django-environ
statistics
nltk
bs4
google-cloud-bigquery
opensearch-py
openai
msal
PyJWT